+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Cors origins variable to settings
 * Current user profile endpoint
+* My polls endpoint
 
 ## v0.4.0 - 22.10.2020